requirements.txt
httpx==0.26.0
rich==13.7.0
google-generativeai==0.3.0
python-dotenv==1.0.0
aiohttp==3.9.1
argparse>=1.4.0 
